requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
python-multipart==0.0.22
pydantic==2.12.5

# Original dependencies from agent.py & rag.py
langchain==1.2.11
langchain-google-vertexai==3.2.2
langchain-google-genai
langchain-core==1.2.18
langchain-community==0.4.1
langchain-classic==1.0.2
langchain-text-splitters==1.1.1
google-cloud-aiplatform==1.141.0
google-cloud-storage==3.9.0
pypdf==6.7.5
requests>=2.31.0
structlog>=24.0.0
pydantic-settings>=2.0.0
langgraph

